requirements.txt
django
django-debug-toolbar
qrcode==5.1
django-location-field
pillow
djangorestframework==3.7.7
dynamic-rest
django-rest-swagger
markdown
django-filter
https://github.com/darklow/django-suit/tarball/v2
dj-database-url
whitenoise
django-annoying
pyfcm
easy-thumbnails
unicodecsv
django-import-export
django-widget-tweaks
fcm-django
django-storages
#boto3
django-haystack
drf-haystack
whoosh
mysqlclient
django-mysql
dj_database_url
gunicorn
django_extensions